модальное окно

+++ api/app/galleries.js
@@ -34,11 +34,7 @@ router.get('/', async (req, res, next) => {
 
 router.post('/', auth, upload.single('image'), async (req, res, next) => {
     try {
-        if (!req.body.title || !req.body.image) {
-            return res.status(400).send({message: 'Title and image are required'});
-        }
-
-        const galleryData = {
+            const galleryData = {
             user: req.user._id,
             title: req.body.title,
             image: req.file.filename,
@@ -56,10 +52,14 @@ router.post('/', auth, upload.single('image'), async (req, res, next) => {
 
 router.delete('/:id', auth, async (req, res, next) => {
     try {
-        const gallery = await Gallery.findById(req.params.id);
-        await Gallery.deleteOne(gallery);
+        if (req.body.user === req.user._id.toString()) {
+            const gallery = await Gallery.findById(req.params.id);
+            await Gallery.deleteOne(gallery);
 
-        return res.send({message: 'Deleted!'});
+            return res.send({message: 'Deleted!'});
+        }
+        return res.send({message: 'No access!'});
+
     } catch (e) {
         next(e);
     }
